requirements.txt
google-generativeai==0.8.6
pandas==2.2.3
plotly==6.7.0
python-dotenv==1.2.1
requests==2.32.5
streamlit==1.50.0
streamlit-lottie==0.0.5
pytest>=8.0.0
pytest-cov>=5.0.0
